✨ feat: add brevity soft-rule to system prompt base rules
Cap the primary statement field at 3 sentences (ja/en, scope-parallel)
so phase-rich scenarios stay readable in playback. inner_thought is
intentionally unconstrained. Wording to be A/B-validated on the
harness (see #877). PromptBuilderTests split into a sibling-file
extension to stay under type_body_length.
